build: introduced smeka build system
Changed the build system to smeka.

Further the build system requires that the pre-processor
does not remove // concatenations. This happens if one
cannot get ones pre-processor to disregard comments.
This is only a problem in older cpp. However, it is
easy to come by as we already rely on sed.

Signed-off-by: Nick Papior <nickpapior@gmail.com>