bug: fixed parallel IO
Reverted a lot of defaults in ncdf_init.
The internal structure of the netcdf-fortran code is not
conforming to NF90_PNETCDF mode and thus we should retain
the original defaults.
This means an ABI break with prior NCDF versions as
certain default modes are not set.

Essentially if one only passed the NF90_MPIIO mode, and
PNetCDF is compiled in, one gets that support. Else,
one gets NetCDF4 support.

Signed-off-by: Nick Papior <nickpapior@gmail.com>