bld: added fdict.inc containing versions
The fdict.inc is a header only file which adds _FDICT_VERSION_ information
to enable developers to utilize several different versions of the library
in one single code.

This is a long-standing issue in fortran codes since each project has
to have this maintained individually. Here we try and streamline the
effort by using preprocessor lines to expose variables for devs.

Note that we prefix the variables with _ to ensure they do not
interfere with regular fortran code (as _ is not allowed as the
first character in a variable/function name).

Signed-off-by: Nick Papior <nickpapior@gmail.com>